chore : AdminController, ChatGPTController, InquiryController 에러 처리 재구성

+++ src/main/java/com/project/thejapenproject/controller/InquiryController.java
@@ -2,6 +2,8 @@ package com.project.thejapenproject.controller;
 
 import com.project.thejapenproject.command.InquiryVO;
 import com.project.thejapenproject.command.UserVO;
+import com.project.thejapenproject.command.exception.RequestParameterException;
+import com.project.thejapenproject.command.exception.code.ErrorCode;
 import com.project.thejapenproject.common.annotation.NoneAuth;
 import com.project.thejapenproject.common.annotation.NoneCheckToken;
 import com.project.thejapenproject.inquiry.service.InquiryService;
@@ -33,6 +35,9 @@ public class InquiryController {
     @NoneAuth
     @PostMapping("/insertData")
     public ResponseEntity<String> insertDate(@RequestBody InquiryVO vo) {
+        if (Objects.isNull(vo)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
         int result = inquiryService.insertData(vo);
         if (result == 1) return ResponseEntity.ok("성공");
         else return ResponseEntity.status(HttpStatus.BAD_GATEWAY).body("실패");
@@ -57,9 +62,15 @@ public class InquiryController {
     @NoneAuth
     @PostMapping("/checkPassword")
     public ResponseEntity<Boolean> checkPassword(@RequestBody InquiryVO vo) {
+        if (Objects.isNull(vo)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
         try {
-            if (inquiryService.checkPassword(vo)) return ResponseEntity.ok(true);
-            else return ResponseEntity.ok(false);
+            if (inquiryService.checkPassword(vo)) {
+                return ResponseEntity.ok(true);
+            } else {
+                return ResponseEntity.ok(false);
+            }
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
@@ -73,6 +84,9 @@ public class InquiryController {
     @NoneAuth
     @GetMapping("/getDetails")
     public ResponseEntity<Object> getDetails(@Param("inquiry_num") int inquiry_num) {
+        if (Objects.isNull(inquiry_num)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
         InquiryVO inquiryVO = inquiryService.getDetails(inquiry_num);
         if (inquiryVO == null) return ResponseEntity.ok(HttpStatus.NOT_FOUND);
         return ResponseEntity.ok(inquiryVO);
@@ -86,6 +100,9 @@ public class InquiryController {
     @NoneAuth
     @GetMapping("/deleteData")
     public ResponseEntity<String> deleteData(@Param("inquiry_num") int inquiry_num) throws Exception {
+        if (Objects.isNull(inquiry_num)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
         int result = inquiryService.deleteData(inquiry_num);
         if (result != 1) {
             throw new Exception();
@@ -101,23 +118,29 @@ public class InquiryController {
     @NoneAuth
     @GetMapping("/searchInquiry")
     public ResponseEntity<Object> searchInquiry(@Param("word") String word) {
+        if (Objects.isNull(word)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
         ArrayList<InquiryVO> inquiryList = inquiryService.searchInquiry(word);
         if (inquiryList.size() == 0) return ResponseEntity.ok(HttpStatus.NOT_FOUND);
         return ResponseEntity.ok(inquiryList);
     }
+
     /**
      * 문의내역 comment 작성 API
+     *
      * @Param InquiryVO : 검색 키워드를 통해 조회
-     * **/
+     **/
     @NoneCheckToken
     @PostMapping("/addComment")
     public ResponseEntity<String> addComment(@RequestBody InquiryVO vo) throws Exception {
-        try {
-            int result = inquiryService.addComment(vo);
-            if(result != 1 ) throw new Exception();
-            return ResponseEntity.ok("성공");
-        } catch (Exception e) {
-            throw new RuntimeException(e);
+        if (Objects.isNull(vo)) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
         }
+        int result = inquiryService.addComment(vo);
+        if (result != 1) {
+            throw new RequestParameterException(ErrorCode.WRONG_PARAM);
+        }
+        return ResponseEntity.ok("성공");
     }
 }
